Bigint for primary_key on rails 5.0
